demo_app/requirements.txt
Django==1.9.10
django-crispy-forms==1.6.1
django-formtools==1.0
django-reversion==2.0.6
httplib2==0.9.2
requests==2.11.1
wheel==0.24.0
XlsxWriter==0.9.3
xlwt==1.1.2
djangorestframework